requirements.txt
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
clearbit==0.1.7
decorator==4.4.0
Django==2.2.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
mock==3.0.5
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyhunter==1.2
PyJWT==1.7.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
